refactor(calendar): rename CalendarModule.tsx to kebab-case
Aligns with the codebase naming convention where all 68 other files
use kebab-case filenames.
