+++ src/components/calendar/index.ts
@@ -1,7 +1,7 @@
 // Re-export all calendar views
 
 // Export the main calendar module
-export { CalendarModule } from "./CalendarModule";
+export { CalendarModule } from "./calendar-module";
 
 // Re-export all calendar components
 export * from "./components";
